OS-73 : Review add uuid and translation
